Crate zip

An ergonomic API for reading and writing ZIP files.

The current implementation is based on PKWARE’s APPNOTE.TXT v6.3.9

Re-exports

pub use crate::read::ZipArchive;

pub use crate::write::ZipWriter;

Modules

read

Types for reading ZIP archives

result

Error types that can be emitted from this library

write

Types for creating ZIP archives

Structs

DateTime

A DateTime field to be used for storing timestamps in a zip file

Enums

CompressionMethod

Identifies the storage format used to compress a file within a ZIP archive.