zip

Module read

Types for reading ZIP archives

Structs

AesInfo

Holds the AES information of a file in the zip archive

Config

Configuration for reading ZIP archives.

ZipArchive

ZIP archive reader

ZipFile

A struct for reading a zip file

ZipFileSeek

A struct for reading and seeking a zip file

Enums

ArchiveOffset

The offset of the start of the archive from the beginning of the reader.

Traits

HasZipMetadata

A trait for exposing file metadata inside the zip.

RootDirFilter

A filter that determines whether an entry should be ignored when searching for the root directory of a Zip archive.

Functions

read_zipfile_from_stream

Read ZipFile structures from a non-seekable reader.

root_dir_common_filter

Common filters when finding the root directory of a Zip archive.