zip::read

Trait RootDirFilter

pub trait RootDirFilter: Fn(&Path) -> bool { }

A filter that determines whether an entry should be ignored when searching for the root directory of a Zip archive.

Returns true if the entry should be considered, and false if it should be ignored.

See root_dir_common_filter for a sensible default filter.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".