Struct zip::ZipFile

pub struct ZipFile {
    pub encrypted: bool,
    pub compression_method: CompressionMethod,
    pub last_modified_time: Tm,
    pub crc32: u32,
    pub compressed_size: u64,
    pub uncompressed_size: u64,
    pub file_name: String,
    pub file_comment: String,
    pub header_start: u64,
    pub data_start: u64,
}

Structure representing a ZIP file.

Fields

encrypted

True if the file is encrypted.

compression_method

Compression method used to store the file

last_modified_time

Last modified time. This will only have a 2 second precision.

crc32

CRC32 checksum

compressed_size

Size of the file in the ZIP

uncompressed_size

Size of the file when extracted

file_name

Name of the file

file_comment

File comment

header_start

Specifies where the local header of the file starts

data_start

Specifies where the compressed data of the file starts