Struct zip::ZipReader

pub struct ZipReader<T> {
    // some fields omitted
}

Wrapper for reading the contents of a ZIP file.

fn doit() -> std::io::IoResult<()>
{
    // For demonstration purposes we read from an empty buffer.
    // Normally a File object would be used.
    let buf = [0u8, 128];
    let mut reader = std::io::BufReader::new(&buf);

    let zip = try!(zip::ZipReader::new(reader));

    for file in zip.files()
    {
        println!("Filename: {}", file.file_name);
        let mut file_reader = try!(zip.read_file(file));
        let first_byte = try!(file_reader.read_byte());
        println!("{}", first_byte);
    }
    Ok(())
}

println!("Result: {}", doit());

Methods

impl<T: Reader + Seek> ZipReader<T>

fn new(reader: T) -> IoResult<ZipReader<T>>

Opens a ZIP file and parses the content headers.

fn files(&self) -> Items<ZipFile>

An iterator over the information of all contained files.

fn read_file(&self, file: &ZipFile) -> IoResult<Box<Reader>>

Gets a reader for a contained zipfile.

Possible errors:

ResourceUnavailable: when another reader returned from this function is still active

OtherIoError: if the file is encrypted or has an unsupported compression

fn unwrap(self) -> T

Unwrap and return the inner reader object

The position of the reader is undefined.