Struct zip::ZipWriter

pub struct ZipWriter<W> {
    // some fields omitted
}

Generator for ZIP files.

fn doit() -> std::io::IoResult<()>
{
    // For this example we write to a buffer, but normally you should use a File
    let mut buf = [0u8, ..65536];
    let w = std::io::BufWriter::new(&mut buf);
    let mut zip = zip::ZipWriter::new(w);

    try!(zip.start_file("hello_world.txt", zip::compression::Stored));
    try!(zip.write(b"Hello, World!"));

    // Optionally finish the zip. (this is also done on drop)
    try!(zip.finish());

    Ok(())
}

println!("Result: {}", doit());

Methods

impl<W: Writer + Seek> ZipWriter<W>

fn new(inner: W) -> ZipWriter<W>

Initializes the ZipWriter.

Before writing to this object, the start_file command should be called.

fn start_file(&mut self, name: &str, compression: CompressionMethod) -> IoResult<()>

Start a new file for with the requested compression method.

fn finish(self) -> IoResult<W>

Finish the last file and write all other zip-structures

This will return the writer, but one should normally not append any data to the end of the file. Note that the zipfile will also be finished on drop.