zip::unstable::stream

Trait ZipStreamVisitor

pub trait ZipStreamVisitor {
    // Required methods
    fn visit_file<R: Read>(
        &mut self,
        file: &mut ZipFile<'_, R>,
    ) -> ZipResult<()>;
    fn visit_additional_metadata(
        &mut self,
        metadata: &ZipStreamFileMetadata,
    ) -> ZipResult<()>;
}

Visitor for ZipStreamReader

Required Methods

fn visit_file<R: Read>(&mut self, file: &mut ZipFile<'_, R>) -> ZipResult<()>

file - contains the content of the file and most of the metadata, except:

comment: set to an empty string

data_start: set to 0

external_attributes: unix_mode(): will return None

fn visit_additional_metadata( &mut self, metadata: &ZipStreamFileMetadata, ) -> ZipResult<()>

This function is guranteed to be called after all visit_files.

metadata - Provides missing metadata in visit_file.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.