Trait zip::unstable::write::FileOptionsExt

pub trait FileOptionsExt {
    // Required method
    fn with_deprecated_encryption(self, password: &[u8]) -> Self;
}

Unstable methods for FileOptions.

Required Methods

fn with_deprecated_encryption(self, password: &[u8]) -> Self

Write the file with the given password using the deprecated ZipCrypto algorithm.

This is not recommended for new archives, as ZipCrypto is not secure.

Object Safety

This trait is not object safe.