Crate zip

A library for reading and writing ZIP archives. ZIP is a format designed for cross-platform file “archiving”. That is, storing a collection of files in a single datastream to make them easier to share between computers. Additionally, ZIP is able to compress and encrypt files in its archives.

The current implementation is based on PKWARE’s APPNOTE.TXT v6.3.9

zip has support for the most common ZIP archives found in common use. However, in special cases, there are some zip archives that are difficult to read or write.

This is a list of supported features:

Reading	Writing

Deflate	✅ ->	✅ ->

Modules

read

Types for reading ZIP archives

result

Error types that can be emitted from this library

write

Types for creating ZIP archives

Structs

DateTime

Representation of a moment in time.

ZipArchive

ZIP archive reader

ZipWriter

ZIP archive generator

Enums

CompressionMethod

Identifies the storage format used to compress a file within a ZIP archive.

Constants

SUPPORTED_COMPRESSION_METHODS

The compression methods which have been implemented.