Crate zip

Iterator zippers

This crate provides "zippers" as tuple structs: Zip2, Zip3, etc. These structs accept iterators as fields, and provides iteration over the "zipped" elements.

The functionality is similar to the Iterator::zip method. The difference is that the zip method always produces an iterator that yields two-element tuples. If you want to zip 3 iterators, you'll write xs.zip(ys).zip(zs) which returns an iterator that yields elements of type ((X, Y), Z). On the other hand, with this library you can write Zip3(xs, ys, zs) which is an iterator that yields elements of type (X, Y, Z).

use zip::Zip3;

let chars = ['a', 'b', 'c'];
let mut v = vec![0, 1, 2];
for (&c, i, &mut j) in Zip3(chars.iter(), 0i32..5, v.iter_mut()) {
    assert!(i < 3);
    assert_eq!(i, j);
}

Macros

zip!
This macro emulates an "any-arity" free function that zips iterators

Structs

Zip2
Two-iterator zipper

Zip3
Three-iterator zipper

Zip4
Four-iterator zipper

Zip5
Five-iterator zipper