zip::read

Function read_zipfile_from_stream

pub fn read_zipfile_from_stream<R: Read>(
    reader: &mut R,
) -> ZipResult<Option<ZipFile<'_, R>>>

Read ZipFile structures from a non-seekable reader.

This is an alternative method to read a zip file. If possible, use the ZipArchive functions as some information will be missing when reading this manner.

Reads a file header from the start of the stream. Will return Ok(Some(..)) if a file is present at the start of the stream. Returns Ok(None) if the start of the central directory is encountered. No more files should be read after this.

The Drop implementation of ZipFile ensures that the reader will be correctly positioned after the structure is done.

Missing fields are:

comment: set to an empty string

data_start: set to 0

external_attributes: unix_mode(): will return None