Struct zip::read::ZipArchive

pub struct ZipArchive<R> { /* private fields */ }

ZIP archive reader

use std::io::prelude::*;
fn list_zip_contents(reader: impl Read + Seek) -> zip::result::ZipResult<()> {
    let mut zip = zip::ZipArchive::new(reader)?;

    for i in 0..zip.len() {
        let mut file = zip.by_index(i)?;
        println!("Filename: {}", file.name());
        std::io::copy(&mut file, &mut std::io::stdout());
    }

    Ok(())
}

Implementations

impl<R: Read + Seek> ZipArchive<R>

pub fn new(reader: R) -> ZipResult<ZipArchive<R>>

Read a ZIP archive, collecting the files it contains

This uses the central directory record of the ZIP file, and ignores local file headers

pub fn extract<P: AsRef<Path>>(&mut self, directory: P) -> ZipResult<()>

Extract a Zip archive into a directory, overwriting files if they already exist. Paths are sanitized with ZipFile::enclosed_name.

Extraction is not atomic; If an error is encountered, some of the files may be left on disk.

pub fn len(&self) -> usize

Number of files contained in this zip.

pub fn is_empty(&self) -> bool

Whether this zip archive contains no files

pub fn offset(&self) -> u64

Get the offset from the beginning of the underlying reader that this zip begins at, in bytes.

Normally this value is zero, but if the zip has arbitrary data prepended to it, then this value will be the size of that prepended data.

pub fn comment(&self) -> &[u8]

Get the comment of the zip archive.

pub fn file_names(&self) -> impl Iterator<Item = &str>

Returns an iterator over all the file and directory names in this archive.

pub fn by_name_decrypt<'a>(
&'a mut self,
name: &str,
password: &[u8]
) -> ZipResult<Result<ZipFile<'a>, InvalidPassword>>

Search for a file entry by name, decrypt with given password

pub fn by_name<'a>(&'a mut self, name: &str) -> ZipResult<ZipFile<'a>>

Search for a file entry by name

pub fn by_index_decrypt<'a>(
&'a mut self,
file_number: usize,
password: &[u8]
) -> ZipResult<Result<ZipFile<'a>, InvalidPassword>>

Get a contained file by index, decrypt with given password

pub fn by_index<'a>(&'a mut self, file_number: usize) -> ZipResult<ZipFile<'a>>

Get a contained file by index

pub fn by_index_raw<'a>(
&'a mut self,
file_number: usize
) -> ZipResult<ZipFile<'a>>

Get a contained file by index without decompressing it

pub fn into_inner(self) -> R

Unwrap and return the inner reader object

The position of the reader is undefined.