zip::read

Struct ZipFile

pub struct ZipFile<'a, R: Read> { /* private fields */ }

A struct for reading a zip file

Implementations

impl<'a, R: Read> ZipFile<'a, R>

Methods for retrieving information on zip files

pub fn version_made_by(&self) -> (u8, u8)

Get the version of the file

pub fn name(&self) -> &str

Get the name of the file

Warnings

It is dangerous to use this name directly when extracting an archive. It may contain an absolute path (/etc/shadow), or break out of the current directory (../runtime). Carelessly writing to these paths allows an attacker to craft a ZIP archive that will overwrite critical files.

You can use the ZipFile::enclosed_name method to validate the name as a safe path.

pub fn name_raw(&self) -> &[u8]

Get the name of the file, in the raw (internal) byte representation.

The encoding of this data is currently undefined.

pub fn sanitized_name(&self) -> PathBuf

๐Deprecated since 0.5.7: by stripping ..s from the path, the meaning of paths can change. mangled_name can be used if this behaviour is desirable

Get the name of the file in a sanitized form. It truncates the name to the first NULL byte, removes a leading โ/โ and removes โ..โ parts.

pub fn mangled_name(&self) -> PathBuf

Rewrite the path, ignoring any path components with special meaning.

Absolute paths are made relative

ParentDirs are ignored

Truncates the filename at a NULL byte

This is appropriate if you need to be able to extract something from any archive, but will easily misrepresent trivial paths like foo/../bar as foo/bar (instead of bar). Because of this, ZipFile::enclosed_name is the better option in most scenarios.

pub fn enclosed_name(&self) -> Option<PathBuf>

Ensure the file path is safe to use as a Path.

It canโt contain NULL bytes

It canโt resolve to a path outside the current directory

foo/../bar is fine, foo/../../bar is not.

It canโt be an absolute path

This will read well-formed ZIP files correctly, and is resistant to path-based exploits. It is recommended over ZipFile::mangled_name.

pub fn comment(&self) -> &str

Get the comment of the file

pub fn compression(&self) -> CompressionMethod

Get the compression method used to store the file

pub fn encrypted(&self) -> bool

Get if the files is encrypted or not

pub fn compressed_size(&self) -> u64

Get the size of the file, in bytes, in the archive

pub fn size(&self) -> u64

Get the size of the file, in bytes, when uncompressed

pub fn last_modified(&self) -> Option<DateTime>

Get the time the file was last modified

pub fn is_dir(&self) -> bool

Returns whether the file is actually a directory

pub fn is_symlink(&self) -> bool

Returns whether the file is actually a symbolic link

pub fn is_file(&self) -> bool

Returns whether the file is a normal file (i.e. not a directory or symlink)

pub fn unix_mode(&self) -> Option<u32>

Get unix mode for the file

pub fn crc32(&self) -> u32

Get the CRC32 hash of the original file

pub fn extra_data(&self) -> Option<&[u8]>

Get the extra data of the zip header for this file

pub fn data_start(&self) -> u64

Get the starting offset of the data of the compressed file

pub fn header_start(&self) -> u64

Get the starting offset of the zip header for this file

pub fn central_header_start(&self) -> u64

Get the starting offset of the zip header in the central directory for this file

pub fn options(&self) -> SimpleFileOptions

Get the SimpleFileOptions that would be used to write this file to a new zip archive.

impl<R: Read> ZipFile<'_, R>

Methods for retrieving information on zip files

pub fn extra_data_fields(&self) -> impl Iterator<Item = &ExtraField>

iterate through all extra fields