Struct zip::read::ZipFile

pub struct ZipFile<'a> { /* fields omitted */ }

A struct for reading a zip file

Methods

impl<'a> ZipFile<'a>

Methods for retrieving information on zip files

pub fn version_made_by(&self) -> (u8, u8)

Get the version of the file

pub fn name(&self) -> &str

Get the name of the file

pub fn name_raw(&self) -> &[u8]

Get the name of the file, in the raw (internal) byte representation.

pub fn sanitized_name(&self) -> PathBuf

Get the name of the file in a sanitized form. It truncates the name to the first NULL byte, removes a leading '/' and removes '..' parts.

pub fn comment(&self) -> &str

Get the comment of the file

pub fn compression(&self) -> CompressionMethod

Get the compression method used to store the file

pub fn compressed_size(&self) -> u64

Get the size of the file in the archive

pub fn size(&self) -> u64

Get the size of the file when uncompressed

pub fn last_modified(&self) -> DateTime

Get the time the file was last modified

pub fn unix_mode(&self) -> Option<u32>

Get unix mode for the file

pub fn crc32(&self) -> u32

Get the CRC32 hash of the original file

pub fn data_start(&self) -> u64

Get the starting offset of the data of the compressed file