zip
Struct DateTime

pub struct DateTime { /* private fields */ }

Representation of a moment in time.

Zip files use an old format from DOS to store timestamps, with its own set of peculiarities. For example, it has a resolution of 2 seconds!

A DateTime can be stored directly in a zipfile with FileOptions::last_modified_time, or read from one with ZipFile::last_modified.

Warning

Because there is no timezone associated with the DateTime, they should ideally only be used for user-facing descriptions.

Modern zip files store more precise timestamps; see crate::extra_fields::ExtendedTimestamp for details.

Implementations

impl DateTime

pub fn default_for_write() -> Self

Returns the current time if possible, otherwise the default of 1980-01-01.

impl DateTime

pub const unsafe fn from_msdos_unchecked( datepart: u16, timepart: u16, ) -> DateTime

Converts an msdos (u16, u16) pair to a DateTime object

Safety

The caller must ensure the date and time are valid.

pub fn try_from_msdos( datepart: u16, timepart: u16, ) -> Result<DateTime, DateTimeRangeError>

Converts an msdos (u16, u16) pair to a DateTime object if it represents a valid date and time.

pub fn from_date_and_time( year: u16, month: u8, day: u8, hour: u8, minute: u8, second: u8, ) -> Result<DateTime, DateTimeRangeError>

Constructs a DateTime from a specific date and time

The bounds are:

year: [1980, 2107]

month: [1, 12]

day: [1, 28..=31]

hour: [0, 23]

minute: [0, 59]

second: [0, 58]

pub fn is_valid(&self) -> bool

Indicates whether this date and time can be written to a zip archive.

pub fn from_time(dt: OffsetDateTime) -> Result<DateTime, DateTimeRangeError>

๐Deprecated since 0.6.4: use DateTime::try_from() instead

Converts a OffsetDateTime object to a DateTime

Returns Err when this object is out of bounds

pub const fn timepart(&self) -> u16

Gets the time portion of this datetime in the msdos representation

pub const fn datepart(&self) -> u16

Gets the date portion of this datetime in the msdos representation

pub fn to_time(&self) -> Result<OffsetDateTime, ComponentRange>

๐Deprecated since 1.3.1: use OffsetDateTime::try_from() instead

Converts the DateTime to a OffsetDateTime structure

pub const fn year(&self) -> u16

Get the year. There is no epoch, i.e. 2018 will be returned as 2018.

pub const fn month(&self) -> u8

Get the month, where 1 = january and 12 = december

Warning

When read from a zip file, this may not be a reasonable value

pub const fn day(&self) -> u8

Get the day

Warning

When read from a zip file, this may not be a reasonable value

pub const fn hour(&self) -> u8

Get the hour

Warning

When read from a zip file, this may not be a reasonable value

pub const fn minute(&self) -> u8

Get the minute

Warning

When read from a zip file, this may not be a reasonable value

pub const fn second(&self) -> u8

Get the second

Warning

When read from a zip file, this may not be a reasonable value