Module zip::write

Types for creating ZIP archives

Structs

ExtendedFileOptions

The Extension for Extra Data and Central Extra Data

FileOptions

Metadata for a file to be written

ZipWriter

ZIP archive generator

Traits

FileOptionExtension

File options Extensions

Type Aliases

FullFileOptions

Adds Extra Data and Central Extra Data. It does not implement copy.

SimpleFileOptions

Simple File Options. Can be copied and good for simple writing zip files