Struct zip::write::ZipWriter

pub struct ZipWriter<W: Write + Seek> {
    // some fields omitted
}

Generator for ZIP files.

fn doit() -> zip::result::ZipResult<()>
{
    use std::io::Write;

    // For this example we write to a buffer, but normally you should use a File
    let mut buf: &mut [u8] = &mut [0u8; 65536];
    let mut w = std::io::Cursor::new(buf);
    let mut zip = zip::ZipWriter::new(w);

    try!(zip.start_file("hello_world.txt", zip::CompressionMethod::Stored));
    try!(zip.write(b"Hello, World!"));

    // Optionally finish the zip. (this is also done on drop)
    try!(zip.finish());

    Ok(())
}

println!("Result: {:?}", doit());

Methods

impl<W: Write + Seek> ZipWriter<W>

fn new(inner: W) -> ZipWriter<W>

Initializes the ZipWriter.

Before writing to this object, the start_file command should be called.

fn start_file<S>(&mut self, name: S, compression: CompressionMethod) -> ZipResult<()> where S: Into<String>

Start a new file for with the requested compression method.

fn finish(self) -> ZipResult<W>

Finish the last file and write all other zip-structures

This will return the writer, but one should normally not append any data to the end of the file.
Note that the zipfile will also be finished on drop.