Struct zip::write::ZipWriter

pub struct ZipWriter<W: Write + Seek> { /* fields omitted */ }

Generator for ZIP files.

fn doit() -> zip::result::ZipResult<()>
{
    use std::io::Write;

    // For this example we write to a buffer, but normally you should use a File
    let mut buf: &mut [u8] = &mut [0u8; 65536];
    let mut w = std::io::Cursor::new(buf);
    let mut zip = zip::ZipWriter::new(w);

    let options = zip::write::FileOptions::default().compression_method(zip::CompressionMethod::Stored);
    zip.start_file("hello_world.txt", options)?;
    zip.write(b"Hello, World!")?;

    // Optionally finish the zip. (this is also done on drop)
    zip.finish()?;

    Ok(())
}

println!("Result: {:?}", doit().unwrap());

Methods

impl<W: Write + Seek> ZipWriter<W>

Important traits for ZipWriter<W>
impl<W: Write + Seek> Write for ZipWriter<W>

pub fn new(inner: W) -> ZipWriter<W>

Initializes the ZipWriter.

Before writing to this object, the start_file command should be called.

pub fn start_file<S>(&mut self, name: S, options: FileOptions) -> ZipResult<()> where
S: Into<String>,

Starts a file.

pub fn start_file_from_path(
&mut self,
path: &Path,
options: FileOptions
) -> ZipResult<()>

Starts a file, taking a Path as argument.

This function ensures that the '/' path seperator is used. It also ignores all non 'Normal' Components, such as a starting '/' or '..' and '.'.

pub fn add_directory<S>(
&mut self,
name: S,
options: FileOptions
) -> ZipResult<()> where
S: Into<String>,

Add a directory entry.

You can't write data to the file afterwards.

pub fn add_directory_from_path(
&mut self,
path: &Path,
options: FileOptions
) -> ZipResult<()>

Add a directory entry, taking a Path as argument.

This function ensures that the '/' path seperator is used. It also ignores all non 'Normal' Components, such as a starting '/' or '..' and '.'.

pub fn finish(&mut self) -> ZipResult<W>

Finish the last file and write all other zip-structures

This will return the writer, but one should normally not append any data to the end of the file. Note that the zipfile will also be finished on drop.